convert header stats to a react component
This means that the header stats need only be computed during the json request
